+++ src/core/server/saved_objects/service/lib/search_dsl/search_dsl.ts
@@ -34,6 +34,7 @@ import { IndexMapping } from '../../../mappings';
 import { getQueryParams } from './query_params';
 import { getSortingParams } from './sorting_params';
 import { ISavedObjectTypeRegistry } from '../../../saved_objects_type_registry';
+import { SavedObjectsFindOptions } from '../../../types';
 
 type KueryNode = any;
 
@@ -53,6 +54,8 @@ interface GetSearchDslOptions {
   };
   kueryNode?: KueryNode;
   workspaces?: string[];
+  workspacesSearchOperator?: 'AND' | 'OR';
+  ACLSearchParams?: SavedObjectsFindOptions['ACLSearchParams'];
 }
 
 export function getSearchDsl(
@@ -73,6 +76,8 @@ export function getSearchDsl(
     hasReference,
     kueryNode,
     workspaces,
+    workspacesSearchOperator,
+    ACLSearchParams,
   } = options;
 
   if (!type) {
@@ -96,6 +101,8 @@ export function getSearchDsl(
       hasReference,
       kueryNode,
       workspaces,
+      workspacesSearchOperator,
+      ACLSearchParams,
     }),
     ...getSortingParams(mappings, type, sortField, sortOrder),
   };