+++ src/plugins/data_source_management/public/components/data_source_aggregated_view/data_source_aggregated_view.test.tsx
@@ -3,7 +3,7 @@
  * SPDX-License-Identifier: Apache-2.0
  */
 
-import { ShallowWrapper, shallow } from 'enzyme';
+import { ShallowWrapper, mount, shallow } from 'enzyme';
 import React from 'react';
 import { DataSourceAggregatedView } from './data_source_aggregated_view';
 import { SavedObjectsClientContract } from '../../../../../core/public';
@@ -43,26 +43,29 @@ describe('DataSourceAggregatedView', () => {
   });
 
   it('should render normally with local cluster hidden and all options', () => {
-    component = shallow(
+    const container = mount(
       <DataSourceAggregatedView
         fullWidth={false}
         hideLocalCluster={true}
         savedObjectsClient={client}
         notifications={toasts}
-        displayAllCompatibleDataSources={true}
+        displayAllCompatibleDataSources={false}
+        activeDataSourceIds={[]}
       />
     );
-    expect(component).toMatchSnapshot();
+    expect(container).toMatchSnapshot();
     expect(client.find).toBeCalledWith({
       fields: ['id', 'title', 'auth.type'],
       perPage: 10000,
       type: 'data-source',
     });
     expect(toasts.addWarning).toBeCalledTimes(0);
+    const badge = container.find('EuiNotificationBadge').text();
+    expect(badge).toEqual('0');
   });
 
   it('should render normally with local cluster and actice selections', () => {
-    component = shallow(
+    const container = mount(
       <DataSourceAggregatedView
         fullWidth={false}
         hideLocalCluster={false}
@@ -72,33 +75,37 @@ describe('DataSourceAggregatedView', () => {
         activeDataSourceIds={['test1']}
       />
     );
-    expect(component).toMatchSnapshot();
+    expect(container).toMatchSnapshot();
     expect(client.find).toBeCalledWith({
       fields: ['id', 'title', 'auth.type'],
       perPage: 10000,
       type: 'data-source',
     });
     expect(toasts.addWarning).toBeCalledTimes(0);
+    const badge = container.find('EuiNotificationBadge').text();
+    expect(badge).toEqual('1');
   });
 
   it('should render normally with data source filter', () => {
-    component = shallow(
+    const container = mount(
       <DataSourceAggregatedView
         fullWidth={false}
         hideLocalCluster={false}
         savedObjectsClient={client}
         notifications={toasts}
-        displayAllCompatibleDataSources={false}
+        displayAllCompatibleDataSources={true}
         dataSourceFilter={(ds) => ds.attributes.auth.type !== 'no_auth'}
       />
     );
-    expect(component).toMatchSnapshot();
+    expect(container).toMatchSnapshot();
     expect(client.find).toBeCalledWith({
       fields: ['id', 'title', 'auth.type'],
       perPage: 10000,
       type: 'data-source',
     });
     expect(toasts.addWarning).toBeCalledTimes(0);
+    const badge = container.find('EuiNotificationBadge').text();
+    expect(badge).toEqual('All');
   });
 
   it('should render popup when clicking on info icon', async () => {