+++ src/plugins/data_source_management/public/components/data_source_menu/data_source_menu.tsx
@@ -5,10 +5,9 @@
 
 import React, { ReactElement } from 'react';
 
-import { DataSourceSelectable } from './data_source_selectable';
 import { DataSourceAggregatedView } from '../data_source_aggregated_view';
 import { DataSourceView } from '../data_source_view';
-import { DataSourceMultiSelectable } from '../data_source_multi_selectable/data_source_multi_selectable';
+import { DataSourceMultiSelectable } from '../data_source_multi_selectable';
 import {
   DataSourceAggregatedViewConfig,
   DataSourceComponentType,
@@ -17,16 +16,19 @@ import {
   DataSourceSelectableConfig,
   DataSourceViewConfig,
 } from './types';
+import { DataSourceSelectable } from '../data_source_selectable';
 
 export function DataSourceMenu<T>(props: DataSourceMenuProps<T>): ReactElement | null {
-  const { componentType, componentConfig } = props;
+  const { componentType, componentConfig, uiSettings, hideLocalCluster } = props;
 
   function renderDataSourceView(config: DataSourceViewConfig): ReactElement | null {
-    const { activeOption, fullWidth } = config;
+    const { activeOption, fullWidth, savedObjects, notifications } = config;
     return (
       <DataSourceView
-        selectedOption={activeOption && activeOption.length > 0 ? activeOption : undefined}
         fullWidth={fullWidth}
+        selectedOption={activeOption}
+        savedObjectsClient={savedObjects}
+        notifications={notifications?.toasts}
       />
     );
   }
@@ -34,13 +36,7 @@ export function DataSourceMenu<T>(props: DataSourceMenuProps<T>): ReactElement |
   function renderDataSourceMultiSelectable(
     config: DataSourceMultiSelectableConfig
   ): ReactElement | null {
-    const {
-      fullWidth,
-      hideLocalCluster,
-      savedObjects,
-      notifications,
-      onSelectedDataSources,
-    } = config;
+    const { fullWidth, savedObjects, notifications, onSelectedDataSources } = config;
     return (
       <DataSourceMultiSelectable
         fullWidth={fullWidth}
@@ -57,7 +53,6 @@ export function DataSourceMenu<T>(props: DataSourceMenuProps<T>): ReactElement |
       onSelectedDataSources,
       disabled,
       activeOption,
-      hideLocalCluster,
       fullWidth,
       savedObjects,
       notifications,
@@ -73,6 +68,7 @@ export function DataSourceMenu<T>(props: DataSourceMenuProps<T>): ReactElement |
         dataSourceFilter={dataSourceFilter}
         hideLocalCluster={hideLocalCluster || false}
         fullWidth={fullWidth}
+        uiSettings={uiSettings}
       />
     );
   }
@@ -82,7 +78,6 @@ export function DataSourceMenu<T>(props: DataSourceMenuProps<T>): ReactElement |
   ): ReactElement | null {
     const {
       fullWidth,
-      hideLocalCluster,
       activeDataSourceIds,
       displayAllCompatibleDataSources,
       savedObjects,