+++ src/plugins/data_source_management/public/components/data_source_multi_selectable/data_source_multi_selectable.tsx
@@ -70,6 +70,10 @@ export class DataSourceMultiSelectable extends React.Component<
             ...this.state,
             selectedOptions,
           });
+
+          this.props.onSelectedDataSources(
+            selectedOptions.filter((option) => option.checked === 'on')
+          );
         }
       })
       .catch(() => {