+++ src/plugins/data_source_management/public/components/data_source_selector/create_data_source_selector.test.tsx
@@ -6,10 +6,16 @@ import { createDataSourceSelector } from './create_data_source_selector';
 import { SavedObjectsClientContract } from '../../../../../core/public';
 import { notificationServiceMock } from '../../../../../core/public/mocks';
 import React from 'react';
-import { render } from '@testing-library/react';
+import { getByText, render } from '@testing-library/react';
+import { coreMock } from '../../../../../core/public/mocks';
+import {
+  mockDataSourcePluginSetupWithHideLocalCluster,
+  mockDataSourcePluginSetupWithShowLocalCluster,
+} from '../../mocks';
 
 describe('create data source selector', () => {
   let client: SavedObjectsClientContract;
+  const { uiSettings } = coreMock.createSetup();
   const { toasts } = notificationServiceMock.createStartContract();
 
   beforeEach(() => {
@@ -27,7 +33,10 @@ describe('create data source selector', () => {
       hideLocalCluster: false,
       fullWidth: false,
     };
-    const TestComponent = createDataSourceSelector();
+    const TestComponent = createDataSourceSelector(
+      uiSettings,
+      mockDataSourcePluginSetupWithHideLocalCluster
+    );
     const component = render(<TestComponent {...props} />);
     expect(component).toMatchSnapshot();
     expect(client.find).toBeCalledWith({
@@ -36,5 +45,23 @@ describe('create data source selector', () => {
       type: 'data-source',
     });
     expect(toasts.addWarning).toBeCalledTimes(0);
+  });
+
+  it('should ignore props.hideLocalCluster, and show local cluster when data_source.hideLocalCluster is set to false', () => {
+    const props = {
+      savedObjectsClient: client,
+      notifications: toasts,
+      onSelectedDataSource: jest.fn(),
+      disabled: false,
+      hideLocalCluster: true,
+      fullWidth: false,
+    };
+    const TestComponent = createDataSourceSelector(
+      uiSettings,
+      mockDataSourcePluginSetupWithShowLocalCluster
+    );
+    const component = render(<TestComponent {...props} />);
+    expect(component).toMatchSnapshot();
+    expect(getByText(component.container, 'Local cluster')).toBeInTheDocument();
   });
 });