+++ src/plugins/data_source_management/public/components/utils.ts
@@ -16,6 +16,7 @@ import {
   noAuthCredentialAuthMethod,
 } from '../types';
 import { AuthenticationMethodRegistry } from '../auth_registry';
+import { DataSourceOption } from './data_source_menu/types';
 
 export async function getDataSources(savedObjectsClient: SavedObjectsClientContract) {
   return savedObjectsClient
@@ -76,6 +77,60 @@ export async function setFirstDataSourceAsDefault(
     const datasourceId = listOfDataSources[0].id;
     return await uiSettings.set('defaultDataSource', datasourceId);
   }
+}
+
+export function getFilteredDataSources(
+  dataSources: Array<SavedObject<DataSourceAttributes>>,
+  filter?: (dataSource: SavedObject<DataSourceAttributes>) => boolean
+) {
+  return filter ? dataSources.filter((ds) => filter!(ds)) : dataSources;
+}
+
+export function getDefaultDataSource(
+  dataSources: Array<SavedObject<DataSourceAttributes>>,
+  LocalCluster: DataSourceOption,
+  uiSettings?: IUiSettingsClient,
+  hideLocalCluster?: boolean,
+  defaultOption?: DataSourceOption[]
+) {
+  const defaultOptionId = defaultOption?.[0]?.id;
+  const defaultOptionDataSource = dataSources.find(
+    (dataSource) => dataSource.id === defaultOptionId
+  );
+
+  const defaultDataSourceId = uiSettings?.get('defaultDataSource', null) ?? null;
+  const defaultDataSourceAfterCheck = dataSources.find(
+    (dataSource) => dataSource.id === defaultDataSourceId
+  );
+
+  if (defaultOptionDataSource) {
+    return [
+      {
+        id: defaultOptionDataSource.id,
+        label: defaultOption?.[0]?.label || defaultOptionDataSource.attributes?.title,
+      },
+    ];
+  }
+  if (defaultDataSourceAfterCheck) {
+    return [
+      {
+        id: defaultDataSourceAfterCheck.id,
+        label: defaultDataSourceAfterCheck.attributes?.title || '',
+      },
+    ];
+  }
+  if (!hideLocalCluster) {
+    return [LocalCluster];
+  }
+  if (dataSources.length > 0) {
+    return [
+      {
+        id: dataSources[0].id,
+        label: dataSources[0].attributes.title,
+      },
+    ];
+  }
+  return [];
 }
 
 export async function getDataSourceById(
@@ -148,7 +203,7 @@ export async function testConnection(
   });
 }
 
-export async function fetchDataSourceVersion(
+export async function fetchDataSourceMetaData(
   http: HttpStart,
   { endpoint, auth: { type, credentials } }: DataSourceAttributes,
   dataSourceID?: string
@@ -164,7 +219,7 @@ export async function fetchDataSourceVersion(
     },
   };
 
-  return await http.post(`/internal/data-source-management/fetchDataSourceVersion`, {
+  return await http.post(`/internal/data-source-management/fetchDataSourceMetaData`, {
     body: JSON.stringify(query),
   });
 }