+++ src/plugins/home/public/application/components/tutorial_directory.js
@@ -236,6 +236,7 @@ class TutorialDirectoryUi extends React.Component {
           onSelectedDataSource={this.onSelectedDataSourceChange}
           disabled={!isDataSourceEnabled}
           hideLocalCluster={isLocalClusterHidden}
+          uiSettings={getServices().uiSettings}
         />
       </div>
     ) : null;