+++ src/plugins/workspace/public/plugin.test.ts
@@ -9,6 +9,7 @@ import { workspaceClientMock, WorkspaceClientMock } from './workspace_client.moc
 import { applicationServiceMock, chromeServiceMock, coreMock } from '../../../core/public/mocks';
 import { WorkspacePlugin } from './plugin';
 import { WORKSPACE_FATAL_ERROR_APP_ID, WORKSPACE_OVERVIEW_APP_ID } from '../common/constants';
+import { savedObjectsManagementPluginMock } from '../../saved_objects_management/public/mocks';
 
 describe('Workspace plugin', () => {
   const getSetupMock = () => ({
@@ -21,21 +22,25 @@ describe('Workspace plugin', () => {
   });
   it('#setup', async () => {
     const setupMock = getSetupMock();
+    const savedObjectManagementSetupMock = savedObjectsManagementPluginMock.createSetupContract();
     const workspacePlugin = new WorkspacePlugin();
-    await workspacePlugin.setup(setupMock);
-    expect(setupMock.application.register).toBeCalledTimes(3);
+    await workspacePlugin.setup(setupMock, {
+      savedObjectsManagement: savedObjectManagementSetupMock,
+    });
+    expect(setupMock.application.register).toBeCalledTimes(4);
     expect(WorkspaceClientMock).toBeCalledTimes(1);
+    expect(savedObjectManagementSetupMock.columns.register).toBeCalledTimes(1);
   });
 
   it('#call savedObjectsClient.setCurrentWorkspace when current workspace id changed', async () => {
     const workspacePlugin = new WorkspacePlugin();
     const setupMock = getSetupMock();
     const coreStart = coreMock.createStart();
-    await workspacePlugin.setup(setupMock);
+    await workspacePlugin.setup(setupMock, {});
     workspacePlugin.start(coreStart);
     coreStart.workspaces.currentWorkspaceId$.next('foo');
     expect(coreStart.savedObjects.client.setCurrentWorkspace).toHaveBeenCalledWith('foo');
-    expect(setupMock.application.register).toBeCalledTimes(3);
+    expect(setupMock.application.register).toBeCalledTimes(4);
     expect(WorkspaceClientMock).toBeCalledTimes(1);
     expect(workspaceClientMock.enterWorkspace).toBeCalledTimes(0);
   });
@@ -69,8 +74,8 @@ describe('Workspace plugin', () => {
     });
 
     const workspacePlugin = new WorkspacePlugin();
-    await workspacePlugin.setup(setupMock);
-    expect(setupMock.application.register).toBeCalledTimes(3);
+    await workspacePlugin.setup(setupMock, {});
+    expect(setupMock.application.register).toBeCalledTimes(4);
     expect(WorkspaceClientMock).toBeCalledTimes(1);
     expect(workspaceClientMock.enterWorkspace).toBeCalledWith('workspaceId');
     expect(setupMock.getStartServices).toBeCalledTimes(1);
@@ -123,7 +128,7 @@ describe('Workspace plugin', () => {
     });
 
     const workspacePlugin = new WorkspacePlugin();
-    await workspacePlugin.setup(setupMock);
+    await workspacePlugin.setup(setupMock, {});
     currentAppIdSubscriber?.next(WORKSPACE_FATAL_ERROR_APP_ID);
     expect(applicationStartMock.navigateToApp).toBeCalledWith(WORKSPACE_OVERVIEW_APP_ID);
     windowSpy.mockRestore();
@@ -132,7 +137,7 @@ describe('Workspace plugin', () => {
   it('#setup register workspace dropdown menu when setup', async () => {
     const setupMock = coreMock.createSetup();
     const workspacePlugin = new WorkspacePlugin();
-    await workspacePlugin.setup(setupMock);
+    await workspacePlugin.setup(setupMock, {});
     expect(setupMock.chrome.registerCollapsibleNavHeader).toBeCalledTimes(1);
   });
 });